About Othello.python
This Othello game is for 2 players on a 8x8 board. Black tokens are represented by "X"s while white tokens are represented by "O"s. The black player (player 1) goes first. When the game ends, the program will print out who won as well as the scoreboard. After that, the program will ask the user if they want to play again. If the game is played again, the program will keep track of the previous scoreboard and increment the new win based off the previous scoreboard. If the user does not want to play again, the program ends.